+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Fix list endpoints
 - Add ability to create a image with labels
 - Add lots of missing filter variants and document them
+- `ContainerOptsBuilder::expose` and `ContainerOptsBuilder::publish` now take a strongly typed `PublishPort`
+  as input parameter.
 
 # 0.3.3
 - Fix return type of `Image::inspect`